requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.10
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
django-filter==2.4.0
django-formtools==2.3
django-import-export==2.5.0
django-redis==4.12.1
django-reversion==4.0.0
django-timezone-field==4.1.2
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.19.1
idna==2.10
importlib-metadata==4.8.1
itypes==1.2.0
Jinja2==3.0.2
kombu==4.6.11
ldap3==2.9
MarkupPy==1.14
MarkupSafe==2.0.1
odfpy==1.4.1
openpyxl==3.0.9
pyasn1==0.4.8
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
redis==3.5.3
requests==2.25.1
sentry-sdk==1.1.0
six==1.15.0
South==1.0.2
sqlparse==0.4.2
tablib==3.0.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==1.3.0
xlrd==2.0.1
XlsxWriter==3.0.1
xlwt==1.3.0
zipp==3.6.0
uwsgi==2.0.20